Fix all warnings

+++ src/common/mod.rs
@@ -3,7 +3,7 @@ pub mod token;
 
 pub struct PageOption<T> {
     pub offset: Option<T>,
-    pub limit: u32,
+    pub limit: usize,
     pub desc: bool,
 }
